refactor: scaffold domain layout for adapters
