+++ core/adapters/__init__.py
@@ -1,0 +1,3 @@
+"""Provider adapters used by BUS Core."""
+
+__all__ = ["fs", "drive"]
