+++ core/adapters/drive/__init__.py
@@ -1,0 +1,5 @@
+"""Google Drive adapter package."""
+
+from .provider import GoogleDriveProvider
+
+__all__ = ["GoogleDriveProvider"]
