+++ core/adapters/fs/__init__.py
@@ -1,0 +1,5 @@
+"""Local filesystem adapter package."""
+
+from .provider import LocalFSProvider
+
+__all__ = ["LocalFSProvider"]
