+++ core/adapters/fs/provider.py
