+++ core/audit/__init__.py
@@ -1,0 +1,3 @@
+"""Audit log scaffolding for BUS Core."""
+
+__all__ = []
