+++ core/domain/__init__.py
@@ -1,0 +1,6 @@
+"""Domain-layer helpers for BUS Core."""
+
+from .broker import Broker
+from .catalog import CatalogManager
+
+__all__ = ["Broker", "CatalogManager"]
