+++ core/pipeline/__init__.py
@@ -1,0 +1,3 @@
+"""Execution pipeline scaffolding for BUS Core."""
+
+__all__ = []
