+++ core/runtime/__init__.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from typing import Optional, TYPE_CHECKING
 
 if TYPE_CHECKING:  # pragma: no cover
-    from core.broker.runtime import Broker
+    from core.domain.broker import Broker
 
 
 _BROKER: Optional["Broker"] = None
