+++ core/sandbox/__init__.py
@@ -1,0 +1,3 @@
+"""Sandbox execution environment scaffolding."""
+
+__all__ = []
