+++ core/storage/__init__.py
@@ -1,0 +1,3 @@
+"""Storage layer scaffolding for BUS Core."""
+
+__all__ = []
